+++ documentation/example/app/api_auth/iam.tf
@@ -1,7 +1,6 @@
 module "api_role" {
-  source   = "path/to/modules/iam/role/api_gateway"
-  log_data = module.stage_log.data["prod"]
-  name     = "api.example.com-api"
+  source = "path/to/modules/iam/role/api_gateway"
+  name   = "api.example.com_api"
   policy_attach_arn_map = {
     auth_invoke        = module.lambda.data[local.lambda_name_auth].iam_policy_arn_map.write
     integration_invoke = module.lambda.data[local.lambda_name_integration].iam_policy_arn_map.write